Update SugarView.php to draw module icons
After upgrading a SugarCRM CE 6.4.5 to 6.5.16 and then 6.5.22 I noticed the module icons were no longer present.  This appears to be a bug introduced around line 1286 in which the developer uses the array_shift PHP function to lop off the icon parameter from the _getModuleTitleParams() function.  This shouldn't have happened, because it provides a better UI experience.  If it was intentional, this seems a confusing way to disable this more attractive UI feature.